Update install procedures

+++ doc/developer/sublime_text.sh
@@ -23,14 +23,19 @@ if [[ ! "$SUBLIME_TEXT" = "" ]]; then
         export REMOVE_MAKO=1
         conda install mako -c conda-forge -y
     fi        
-    wget https://raw.githubusercontent.com/StatisKit/StatisKit/master/doc/developper/SCons.sublime-build -O SCons.sublime-build
-    echo "import os" > install.py
-    echo "from mako.template import Template" >> install.py
-    echo "with open(os.environ.get('SUBLIME_TEXT') + '/SCons.sublime-build', 'w') as filehandler:" >> install.py
-    echo "    filehandler.write(Template(filename='SCons.sublime-build').render(CONDA_DIR=os.environ.get('CONDA_DIR'), project_path='${project_path}'))" >> install.py
-    python install.py
-    rm install.py
-    rm SCons.sublime-build
+    if [[ ! "$?" = "0" ]]; then
+        wget https://raw.githubusercontent.com/StatisKit/StatisKit/master/doc/developper/SCons.sublime-build -O SCons.sublime-build
+        echo "import os" > install.py
+        echo "from mako.template import Template" >> install.py
+        echo "with open(os.environ.get('SUBLIME_TEXT') + '/SCons.sublime-build', 'w') as filehandler:" >> install.py
+        echo "    filehandler.write(Template(filename='SCons.sublime-build').render(CONDA_DIR=os.environ.get('CONDA_DIR'), project_path='${project_path}'))" >> install.py
+        python install.py
+        if [[ ! "$?" = "0" ]]; then
+            echo 'Sublime Text '$SUBLIME_TEXT' configuration failed.'
+        fi
+        rm install.py
+        rm SCons.sublime-build
+    fi
     if [[ "$REMOVE_MAKO" = "1" ]]; then
         conda remove mako -y
     fi
